+++ backend/config.py
@@ -1,0 +1,7 @@
+DB_CONFIG = {
+    'dbname': 'merrick_server', 
+    'user': 'merrick',
+    'password': 'upset_stomach',
+    'host': '34.125.49.223',  # or your database server's IP
+    'port': 5432          # default PostgreSQL port
+}
